Align case-study page design with issue/solution pages
- Move the parent "Case study of" callout below the title and badges,
  matching how issues render (id → title → badges → parent).
- Hide the owner/collaborators panel behind a dedicated Contributors
  page and the change log behind a History page, reached via quiet meta
  links at the bottom of the overview — the same pattern issues and
  solutions use. Both are now nested routes under /case-study/[id].
